+++ src/main/java/cleancode/minesweeper/tobe/GameApplication.java
@@ -1,0 +1,11 @@
+package cleancode.minesweeper.tobe;
+
+public class MinesweeperGame {
+
+    public static void main(String[] args) {
+        Minesweeper minesweeper = new Minesweeper();
+        minesweeper.run();
+    }
+
+}
+
