+++ src/main/java/cleancode/minesweeper/tobe/GameBoard.java
@@ -1,0 +1,4 @@
+package cleancode.minesweeper.tobe;
+
+public class GameBoard {
+}
