+++ src/main/java/cleancode/minesweeper/tobe/GameException.java
