+++ src/main/java/cleancode/minesweeper/tobe/Minesweeper.java
@@ -1,0 +1,32 @@
+package cleancode.minesweeper.tobe;
+
+public class MineSweaper {
+    
+    public void run() {
+        showGameStartComments();
+        initializeGame();
+
+        while (true) {
+            try {
+                showBoard();
+
+                if (doesUserWinTheGame()) {
+                    System.out.println("지뢰를 모두 찾았습니다. GAME CLEAR!");
+                    break;
+                }
+                if (doesUserLoseTheGame()) {
+                    System.out.println("지뢰를 밟았습니다. GAME OVER!");
+                    break;
+                }
+
+                String cellInput = getCellInputFromUser();
+                String userActionInput = getUserActionInputFromUser();
+                actOnCell(cellInput, userActionInput);
+            } catch (AppException e) {
+                System.out.println(e.getMessage());
+            } catch (Exception e) {
+                System.out.println("프로그램에 문제가 생겼습니다. ");
+            }
+        }
+    }
+}
