+++ src/main/java/cleancode/minesweeper/tobe/io/ConsoleInputHandler.java
@@ -1,0 +1,12 @@
+package cleancode.minesweeper.tobe.io;
+
+import java.util.Scanner;
+
+public class ConsoleInputHandler {
+
+    public static final Scanner SCANNER = new Scanner(System.in);
+
+    public String getUserInput() {
+        return SCANNER.nextLine();
+    }
+}
